Move compass from SceneView to ChromeView.
(cherry picked from commit 7e9d6b629cee028b24bb0e83e096642a83a79630)
